Backend (#20)
* add authorization (#11)

* move serviceAccountKey from resources to end in docker-compose.yaml (#12)

* Update README.md

add description for backend

* Update README.md

* Update README.md

* add my_events and my_enrolls routes (#13)

* Merge pull request #14

* handle not found exceptions

* handle double enroll or unenroll if not enrolled

* add cors config (#15)

* Backend  add images (#16)

* add images

* save images

* fix dockerfile

* count available seats && return image name when creating && return list of enrolled users for owner (#17)

* backend--metrics (#18)

* Add metrics: graphana, prometheus

* collect logs with loki && add logs && complete metrics && document code

* add pre configured datasources and dashboards && small fix

---------


* add places for events && remove auth from GET /event/list and GET /event/{eventId} (#19)

* backend: handle MissingServletRequestParameterException

---------

+++ backend/src/main/java/com/vladdan16/spda_afisha/backend/service/jpa/JpaEventService.java
@@ -1,17 +1,26 @@
 package com.vladdan16.spda_afisha.backend.service.jpa;
 
+import com.vladdan16.spda_afisha.backend.domain.exceptions.ForbiddenException;
+import com.vladdan16.spda_afisha.backend.domain.exceptions.NotFoundException;
 import com.vladdan16.spda_afisha.backend.domain.models.EventType;
 import com.vladdan16.spda_afisha.backend.domain.models.Event;
 import com.vladdan16.spda_afisha.backend.domain.repositories.EventRepository;
 import com.vladdan16.spda_afisha.backend.dto.responses.events.EventResponse;
 import com.vladdan16.spda_afisha.backend.dto.responses.events.ListEventResponse;
+import com.vladdan16.spda_afisha.backend.dto.responses.events.ListOwnerEventResponse;
+import com.vladdan16.spda_afisha.backend.dto.responses.events.OwnerEventResponse;
+import com.vladdan16.spda_afisha.backend.dto.responses.users.UserResponse;
 import com.vladdan16.spda_afisha.backend.service.EventService;
+import com.vladdan16.spda_afisha.backend.service.ImageService;
 import jakarta.transaction.Transactional;
 import jakarta.validation.constraints.NotNull;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.web.multipart.MultipartFile;
 
 import java.sql.Timestamp;
+import java.util.HashSet;
+import java.util.List;
 
 
 @Service
@@ -19,23 +28,30 @@ import java.sql.Timestamp;
 @RequiredArgsConstructor
 public class JpaEventService implements EventService {
   private final EventRepository eventRepository;
+  private final ImageService imageService;
 
   @Override
-  public void createEvent(
+  public Long createEvent(
+      String userId,
       String name,
+      String place,
       String description,
       Timestamp startAt,
       Long numberSeats,
       EventType type
   ) {
     var event = Event.builder()
+        .authorId(userId)
         .name(name)
+        .place(place)
         .description(description)
         .startAt(startAt)
         .numberSeats(numberSeats)
+        .availableSeats(numberSeats)
         .type(type)
         .build();
     eventRepository.save(event);
+    return event.getId();
   }
 
   @Override
@@ -46,44 +62,75 @@ public class JpaEventService implements EventService {
         .map((event) -> new EventResponse(
             event.getId(),
             event.getName(),
+            event.getPlace(),
             event.getDescription(),
             event.getStartAt(),
             event.getNumberSeats(),
-            event.getType())
+            event.getAvailableSeats(),
+            event.getType(),
+            event.getImages())
         ).toList());
   }
 
   @Override
-  public void deleteEvent(Long id) {
+  public void deleteEvent(String userId, Long id) {
+    var event = eventRepository.getEventById(id);
+    if (event == null) {
+      throw new NotFoundException("Event not found");
+    }
+
+    if (!event.getAuthorId().equals(userId)) {
+      throw new ForbiddenException("Unable to delete event not created by current user");
+    }
+    imageService.deleteAllImages(event.getImages());
     eventRepository.deleteById(id);
   }
 
   @Override
   public EventResponse getEvent(Long id) {
-    var event = eventRepository.getReferenceById(id);
+    var event = eventRepository.getEventById(id);
+    if (event == null) {
+      throw new NotFoundException("Event not found");
+    }
+
     return new EventResponse(
         event.getId(),
         event.getName(),
+        event.getPlace(),
         event.getDescription(),
         event.getStartAt(),
         event.getNumberSeats(),
-        event.getType()
+        event.getAvailableSeats(),
+        event.getType(),
+        event.getImages()
     );
   }
 
   @Override
   public void updateEvent(
+      String userId,
       @NotNull Long id,
       String name,
+      String place,
       String description,
       Timestamp startAt,
       Long numberSeats,
       EventType type
   ) {
-    var event = eventRepository.getReferenceById(id);
+    var event = eventRepository.getEventById(id);
+    if (event == null) {
+      throw new NotFoundException("Event not found");
+    }
+
+    if (!event.getAuthorId().equals(userId)) {
+      throw new ForbiddenException("Unable to edit event not created by current user");
+    }
 
     if (name != null) {
       event.setName(name);
+    }
+    if (place != null) {
+      event.setPlace(place);
     }
     if (description != null) {
       event.setDescription(description);
@@ -92,10 +139,78 @@ public class JpaEventService implements EventService {
       event.setStartAt(startAt);
     }
     if (numberSeats != null) {
+      event.setAvailableSeats(event.getAvailableSeats() + numberSeats - event.getNumberSeats());
       event.setNumberSeats(numberSeats);
     }
     if (type != null) {
       event.setType(type);
     }
+  }
+
+  @Override
+  public String saveImage(Long eventId, String userId, MultipartFile file) {
+    var event = eventRepository.getEventById(eventId);
+
+    if (event == null) {
+      throw new NotFoundException("Event not found");
+    }
+
+    if (!event.getAuthorId().equals(userId)) {
+      throw new ForbiddenException("Unable to edit event not created by current user");
+    }
+
+    final var imageName = imageService.storeImage(file);
+
+    event.getImages().add(imageName);
+
+    return imageName;
+  }
+
+  @Override
+  public void deleteImages(Long eventId, String userId, List<String> images) {
+    var event = eventRepository.getEventById(eventId);
+
+    if (event == null) {
+      throw new NotFoundException("Event not found");
+    }
+
+    if (!event.getAuthorId().equals(userId)) {
+      throw new ForbiddenException("Unable to edit event not created by current user");
+    }
+
+    final var set = new HashSet<>(images);
+
+    imageService.deleteAllImages(set.stream().toList());
+
+    event.getImages().removeAll(set);
+  }
+
+  @Override
+  public ListOwnerEventResponse listMyEvents(String userId) {
+    return new ListOwnerEventResponse(eventRepository
+        .findByAuthorId(userId)
+        .stream()
+        .map((event) -> {
+              var users = event
+                  .getUsers()
+                  .stream()
+                  .map(user -> new UserResponse(
+                      user.getEmail(),
+                      user.getName(),
+                      user.getSurname())
+                  ).toList();
+              return new OwnerEventResponse(
+                  event.getId(),
+                  event.getName(),
+                  event.getPlace(),
+                  event.getDescription(),
+                  event.getStartAt(),
+                  event.getNumberSeats(),
+                  event.getAvailableSeats(),
+                  event.getType(),
+                  event.getImages(),
+                  users);
+            }
+        ).toList());
   }
 }
